+++ CHANGELOG.md
@@ -74,6 +74,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - Fixed custom Meta attributes removal when transferred to instrument (#615)
   - Corrected iteration over Instrument within list comprehension
   - Removed unused input arguments
+  - Corrects Instrument today, yesterday, and tomorrow methods by implementing
+    datetime.datetime.utcnow
 - Maintenance
   - nose dependency removed from unit tests
   - Specify dtype for empty pandas.Series for forward compatibility